Logging in users using bcrypt and jwt
installed jsonwebtoken module
created a static method findByCredential that allows to fetch a user by its email and compare its hashed password by the one provided when login in using POST /users/login
created an instance method generateAuthToken that allows us to generate a token and save that token to the user document
created a field in the userSchema which is an array of tokens
created the route handler for  POST  /users/login in the src/routers/user.js. The route handle takes in body of email and password  and use the method findByCredential created in src/model/user.js to find a user. It also generate the jwt token and send the user and the token generated
explore how to create a middleware